# api/api.yml
openapi: 3.0.3
info:
  title: chatz API
  version: 0.2.0
  description: >
    Self-hosted AI chat API. Streaming chat with inline UI-spec rendering,
    MCP tool calling, model discovery across OpenAI-compatible upstreams, and
    admin-provisioned auth. Chat streams use Anthropic-style SSE.
servers:
  - url: /v1
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: chatz_session
  schemas:
...
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true
...
    MCPServer:
      type: object
      required: [id, name, transport, enabled, status]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        transport:
          type: string
          enum: [stdio, http]
        command:
          type: string
        url:
          type: string
        enabled:
          type: boolean
        status:
          type: string
          enum: [connecting, connected, failed, disabled]
          description: Live connection state of the server.
        statusError:
          type: string
          description: Error detail, present when status is failed.
        statusReason:
          type: string
          enum: [unreachable, access_denied, not_responding, failed]
          description: Coarse failure classification, present when status is failed.
        toolCount:
          type: integer
          description: Number of tools discovered, present when connected.
        lastConnectionAttemptAt:
          type: string
          format: date-time
          description: UTC time at which the most recent connection attempt started.
        lastSuccessfulConnectionAt:
          type: string
          format: date-time
          description: UTC time at which this server most recently connected successfully.
        lastConnectionFailureAt:
          type: string
          format: date-time
          description: UTC time at which this server most recently failed to connect or disconnected unexpectedly.
        lastConnectionLatencyMs:
          type: integer
          minimum: 0
          description: Elapsed milliseconds for the most recent completed connection attempt.
        lastError:
          type: string
          description: Most recent connection or unexpected-disconnect error detail. Admin-only and never includes decrypted credentials.
        args:
          type: array
          items:
            type: string
          description: Command arguments (stdio transport).
        headers:
          type: object
          additionalProperties:
            type: string
          description: >-
            HTTP headers (http transport). Authorization values are returned as
            '<scheme> [SECRET]' for edit prefill; resubmitting that unchanged
            value preserves the existing sealed credential. Other header values
            are returned as configured. Admin-only response; sealed at rest and
            never logged (the request logger records no bodies).
        env:
          type: object
          additionalProperties:
            type: string
          description: >-
            Environment variables (stdio transport), decrypted for edit prefill.
            Admin-only response; sealed at rest, never logged.
    UpdateMCPServerRequest:
      type: object
      description: >-
        Edit an MCP server. Every field is optional; an omitted field keeps its
        current value. Providing transport-specific fields replaces them.
      properties:
        name:
          type: string
        transport:
          type: string
          enum: [stdio, http]
        command:
          type: string
        args:
          type: array
          items:
            type: string
        url:
          type: string
        headers:
          type: object
          additionalProperties:
            type: string
          description: >-
            HTTP headers to replace. An Authorization value returned as
            '<scheme> [SECRET]' preserves the existing sealed credential when
            sent back unchanged; any other value replaces it.
        env:
          type: object
          additionalProperties:
            type: string
        enabled:
          type: boolean
...
    CreateMCPServerRequest:
      type: object
      required: [name, transport]
      properties:
        name:
          type: string
        transport:
          type: string
          enum: [stdio, http]
        command:
          type: string
        args:
          type: array
          items:
            type: string
        url:
          type: string
        headers:
          type: object
          additionalProperties:
            type: string
        env:
          type: object
          additionalProperties:
            type: string
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Conflict:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
paths:
...
  /mcp/servers:
    get:
      tags: [mcp]
      operationId: listMCPServers
      summary: List configured MCP servers (admin).
      security:
        - sessionCookie: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MCPServer'
    post:
      tags: [mcp]
      operationId: createMCPServer
      summary: Add an MCP server (admin).
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMCPServerRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MCPServer'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
  /mcp/servers/{serverId}:
    delete:
      tags: [mcp]
      operationId: deleteMCPServer
      summary: Remove an MCP server (admin).
      security:
        - sessionCookie: []
      parameters:
        - name: serverId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Deleted
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [mcp]
      operationId: updateMCPServer
      summary: Edit an MCP server and reconnect (admin).
      security:
        - sessionCookie: []
      parameters:
        - name: serverId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateMCPServerRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MCPServer'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
...